cleanup: enable /WX (aka -Werror) with MSVC (#4095)
Fix all the warnings generated by MSVC, most of them were useful. I also
chose to use pragmas to disable warning 4996 (aka you are using an
insecure function) because we should keep that warning enable in as much
of our code as possible. Previously we were disabling it for full files.

Because users may compile with newer versions of the compiler this PR
also includes a CMake option to disable /WX (and -Werror), in case the
newer (or older I suppose) version of the compiler generates warnings we
have not handled.

+++ google/cloud/log.cc
@@ -40,7 +40,8 @@ struct Timestamp {
     gmtime_r(&tt, &tm);
 #endif
     auto const ss = tp - std::chrono::system_clock::from_time_t(tt);
-    nanos = std::chrono::duration_cast<std::chrono::nanoseconds>(ss).count();
+    nanos = static_cast<std::int32_t>(
+        std::chrono::duration_cast<std::chrono::nanoseconds>(ss).count());
   }
   std::tm tm;
   std::int32_t nanos;
